+++ package.json
@@ -6,8 +6,8 @@
   "license": "ISC",
   "repository": "https://github.com/probot/webhooks.git",
   "scripts": {
-    "start": "node index.js",
-    "start-dev": "concurrently \"nodemon --ignore src/ ./index.js\" \"webpack -w\"",
+    "start": "node ./lib/index.js",
+    "start-dev": "concurrently \"nodemon --ignore src/ ./lib/index.js\" \"webpack -w\"",
     "build": "webpack -p",
     "test": "jest && standard",
     "postinstall": "npm run build"
